+++ docs/changelog.rst
@@ -40,7 +40,7 @@ Changelog
 Released: **07-05-2023**
 
 * **07/05/2023** :
-    * Updated docs
+    * Updated docs and plugin code
     * Bug fixes to improve plugin
     * Updated GitHub Actions workflows and noxfile.py